Work around SocketRocket’s angle-bracket #imports
ARTSRDelegateController.h and ARTSRError.h have the following line:

  #import <SocketRocketAblyFork/ARTSRWebSocket.h>

This works fine when compiling these files inside the
SocketRocketAblyFork framework, but when compiling them directly inside
the SDK this path can’t be resolved. So, we create a fake `include`
directory containing a symlink called SocketRocketAblyFork. When we add
this `include` directory to the header search paths, this makes the
import work without needing to modify the SocketRocket code.

This feels a bit nasty and there may be a better way to solve it.